Seqvec_Embeddings_Test
Here is the initial script with the four pdb IDs and their respective sequences. This is going to be the test script, and we are going to run it in a Docker file with the bio_embeddings installed to hopefully see if the output is compatible with the PPI_GNN model.